+++ OpenIM-SDK/src/main/java/io/openim/android/sdk/listener/OnMsgSendCallback.java
@@ -8,7 +8,7 @@ public interface OnMsgSendCallback extends OnBase<String> {
     /**
      * 发送失败
      */
-    void onError(long code, String error);
+    void onError(int code, String error);
 
     /**
      * 上传进度
